Moves fulltext search helper functions to pg_func module
